fix int size issue

+++ ransac.cpp
@@ -121,7 +121,7 @@ bool ht_ransac_best_indices(headtracker_t& ctx, float& mean_error, Mat& rvec_, M
 
     if (ctx.config.debug)
         fprintf(stderr, "ransac failed maxerr=%f zoom-ratio=%f cur=%d\n",
-                ctx.config.ransac_max_inlier_error, ctx.zoom_ratio, object_points.size());
+                ctx.config.ransac_max_inlier_error, ctx.zoom_ratio, (int) object_points.size());
 
     return false;
 }
